refactor: removed duplicate code
