+++ client/src/Result.js
@@ -18,28 +18,10 @@ import Ratings from "react-ratings-declarative/build/ratings";
 import {AdapterDateFns} from '@mui/x-date-pickers/AdapterDateFns';
 import {LocalizationProvider} from '@mui/x-date-pickers/LocalizationProvider';
 import {DatePicker} from '@mui/x-date-pickers/DatePicker';
+import {displayedDate, getGuestRoom, formatDate} from "./GuestRoomConverter";
 
 
 const socket = new WebSocket('ws://localhost:5000')
-
-function padTo2Digits(num) {
-    return num.toString().padStart(2, '0');
-}
-
-function formatDate(date) { // Date -> YYYY-MM-DD
-    return [
-        date.getFullYear(),
-        padTo2Digits(date.getMonth() + 1),
-        padTo2Digits(date.getDate()),
-    ].join('-');
-}
-
-function displayedDate(date) { //YYYY-MM-DD -> MMM(Jul) DD YYYY
-    let temp_date = new Date(date.split('-')[0],
-        date.split('-')[1] - 1,
-        date.split('-')[2])
-    return temp_date.toString().split(" ").slice(1, 4).join(" ");
-}
 
 class Result extends React.Component {
     constructor(props) {
@@ -82,7 +64,7 @@ class Result extends React.Component {
                 date2: new Date(qs.parse(window.location.search).checkout.split('-')[0],
                     qs.parse(window.location.search).checkout.split('-')[1] - 1,
                     qs.parse(window.location.search).checkout.split('-')[2]),
-                guestNo: this.getGuestRoom(qs.parse(window.location.search).guests),
+                guestNo: getGuestRoom(qs.parse(window.location.search).guests),
             });
             fetch("/search" + window.location.search)
                 .then(response => response.json())
@@ -104,31 +86,6 @@ class Result extends React.Component {
         localStorage.setItem("displayCheckout", displayedDate(this.state.queryParams.checkout));
         localStorage.setItem("guestCount", this.state.queryParams.guests);
         console.log(this.state.queryParams)
-    }
-
-    getGuestRoom(guestsParam) {
-        let ls = guestsParam.split('|');
-        let result = [0, 0, 0, 0];
-        for (let i = 0; i < ls.length; i++) {
-            let curr = ls[i];
-            switch (curr) {
-                case "1" :
-                    result[0] += 1;
-                    break;
-                case "2":
-                    result[1] += 1;
-                    break;
-                case "3":
-                    result[2] += 1;
-                    break;
-                case "4":
-                    result[3] += 1;
-                    break;
-                default:
-                    break;
-            }
-        }
-        return result;
     }
 
     PageBtn(i) {
